Cleaned up vendor link request page
